app/requirements.txt
google-generativeai
Pillow
pyqt5
mss
winsdk